requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
bcrypt==4.1.2
boto3==1.34.64
botocore==1.34.64
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.3.0
cryptography==41.0.7
Deprecated==1.2.14
deprecation==2.1.0
dnspython==2.4.2
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.108.0
fastapi-limiter==0.1.6
fastapi_keycloak==1.0.10
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-resources==6.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
jmespath==1.0.1
limits==3.7.0
orjson==3.9.10
packaging==23.2
paramiko==3.4.0
pluggy==1.4.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pymongo==4.6.1
PyNaCl==1.5.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-keycloak==3.7.0
python-multipart==0.0.6
redis==5.1.0b2
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
wrapt==1.16.0